requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

pydantic==2.7.1
email-validator==2.1.1

pymongo==4.7.2
python-dotenv==1.0.1

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-jose==3.3.0

python-multipart==0.0.9

loguru==0.7.2
